+++ src/heap.js
@@ -1,29 +1,25 @@
 /**
- * @datastructures-js/heap
- * @copyright 2020 Eyas Ranjous <eyas.ranjous@gmail.com>
  * @license MIT
- */
-
-const HeapNode = require('./heapNode');
-
-const isNumber = (n) => typeof n === 'number';
-const isNoneEmptyString = (s) => typeof s === 'string' && s.length;
-const isNoneNullObject = (o) => typeof o === 'object' && o !== null;
-const isNoneEmptyArray = (a) => Array.isArray(a) && a.length > 0;
-
-/**
- * @class Heap
+ * @copyright 2020 Eyas Ranjous <eyas.ranjous@gmail.com>
+ *
+ * @class
  * @abstract
  */
 class Heap {
-  constructor(nodes) {
+  /**
+   * Creates a heap instance
+   * @param {array<string|number|object>} nodes
+   * @param {string|number|object} [leaf]
+   * @returns {number}
+   */
+  constructor(nodes, leaf) {
     this._nodes = Array.isArray(nodes) ? nodes : [];
-    this._leaf = null;
+    this._leaf = leaf || null;
   }
 
   /**
+   * Calculates left child's index from a parent's index
    * @private
-   * calculates the left child's index of a parent's index
    * @param {number} parentIndex
    * @returns {number}
    */
@@ -32,8 +28,8 @@ class Heap {
   }
 
   /**
+   * Calculates right child's index from a parent's index
    * @private
-   * calculates the right child's index of a parent's index
    * @param {number} parentIndex
    * @returns {number}
    */
@@ -42,8 +38,8 @@ class Heap {
   }
 
   /**
+   * Calculates parent's index from a child's index
    * @private
-   * calculates a parent's index from a child's index
    * @param {number} parentIndex
    * @returns {number}
    */
@@ -52,16 +48,49 @@ class Heap {
   }
 
   /**
+   * Checks if a parent has a left child
    * @private
-   * gets the last node's index
-   * @returns {number}
+   * @param {number} parentIndex
+   * @returns {boolean}
    */
-  _getLastIndex() {
-    return this._nodes.length - 1;
+  _hasLeftChild(parentIndex) {
+    return this._getLeftChildIndex(parentIndex) < this.size();
   }
 
   /**
-   * swaps two nodes in the heap by their indices
+   * Checks if a parent has a right child
+   * @private
+   * @param {number} parentIndex
+   * @returns {boolean}
+   */
+  _hasRightChild(parentIndex) {
+    return this._getRightChildIndex(parentIndex) < this.size();
+  }
+
+  /**
+   * Checks if a child has a parent
+   * @private
+   * @param {number} childIndex
+   * @returns {boolean}
+   */
+  _hasParent(childIndex) {
+    return this._getParentIndex(childIndex) >= 0;
+  }
+
+  /**
+   * Returns heap node's key
+   * @private
+   * @param {object|number|string} node
+   * @returns {number|string}
+   */
+  _getKey(node) {
+    if (typeof node === 'object') return node.key;
+    return node;
+  }
+
+  /**
+   * Swaps two nodes in the heap
+   * @private
    * @param {number} i
    * @param {number} j
    */
@@ -72,56 +101,134 @@ class Heap {
   }
 
   /**
+   * Compares parent & child nodes
+   * and returns true if they are in right positions
+   *
    * @private
-   * selects the proper child's index to fix the heap
-   * @param {number} parentIndex
-   * @returns {number}
+   * @param {object|number|string} parent
+   * @param {object|number|string} child
+   * @returns {boolean}
    */
-  _compareChildrenOf(parentIndex) {
-    const leftChildIndex = this._getLeftChildIndex(parentIndex);
-    const rightChildIndex = this._getRightChildIndex(parentIndex);
-    const size = this.size();
-
-    if (leftChildIndex >= size && rightChildIndex >= size) return -1;
-    if (leftChildIndex >= size) return rightChildIndex;
-    if (rightChildIndex >= size) return leftChildIndex;
-
-    return this._compareChildren(leftChildIndex, rightChildIndex);
+  _compare(parentNode, childNode) {
+    return this._compareKeys(
+      this._getKey(parentNode),
+      this._getKey(childNode)
+    );
   }
 
   /**
+   * Compares parent & child nodes
+   * and returns true if they are in right positions
+   *
    * @private
-   * bubbles the last inserted node up in the heap
+   * @param {object|number|string} parent
+   * @param {object|number|string} child
+   * @returns {boolean}
    */
-  _heapifyUp() {
-    let childIndex = this._getLastIndex();
-    let parentIndex = this._getParentIndex(childIndex);
+  _compareByIndex(parentIndex, childIndex) {
+    return this._compareKeys(
+      this._getKey(this._nodes[parentIndex]),
+      this._getKey(this._nodes[childIndex])
+    );
+  }
 
-    while (this._shouldSwap(childIndex, parentIndex)) {
-      this._swap(childIndex, parentIndex);
+  /**
+   * Checks if parent and child nodes should be swapped
+   * @private
+   * @param {number} parentIndex
+   * @param {number} childIndex
+   * @returns {boolean}
+   */
+  _shouldSwap(parentIndex, childIndex) {
+    if (parentIndex < 0 || parentIndex >= this.size()) return false;
+    if (childIndex < 0 || childIndex >= this.size()) return false;
+
+    return !this._compareByIndex(parentIndex, childIndex);
+  }
+
+  /**
+   * Bubbles a node from a starting index up in the heap
+   * @param {number} [startingIndex]
+   * @public
+   */
+  heapifyUp(startingIndex = this.size() - 1) {
+    let childIndex = startingIndex;
+    let parentIndex = this._getParentIndex(childIndex);
+    while (this._shouldSwap(parentIndex, childIndex)) {
+      this._swap(parentIndex, childIndex);
       childIndex = parentIndex;
       parentIndex = this._getParentIndex(childIndex);
     }
   }
 
   /**
+   * Compares left and right & children of a parent
    * @private
-   * pushes the replaced root node down in the heap after root's removal
+   * @param {number} parentIndex
+   * @returns {number} - a child's index
    */
-  _heapifyDown() {
-    let parentIndex = 0;
+  _compareChildrenOf(parentIndex) {
+    if (
+      !this._hasLeftChild(parentIndex)
+      && !this._hasRightChild(parentIndex)
+    ) {
+      return -1;
+    }
+
+    const leftChildIndex = this._getLeftChildIndex(parentIndex);
+    const rightChildIndex = this._getRightChildIndex(parentIndex);
+
+    if (!this._hasLeftChild(parentIndex)) {
+      return rightChildIndex;
+    }
+
+    if (!this._hasRightChild(parentIndex)) {
+      return leftChildIndex;
+    }
+
+    return this._compareByIndex(leftChildIndex, rightChildIndex)
+      ? leftChildIndex
+      : rightChildIndex;
+  }
+
+  /**
+   * Pushes a node from a starting index down in the heap
+   * @private
+   */
+  _heapifyDown(startingIndex = 0) {
+    let parentIndex = startingIndex;
     let childIndex = this._compareChildrenOf(parentIndex);
-    while (this._shouldSwap(childIndex, parentIndex)) {
-      this._swap(childIndex, parentIndex);
+    while (this._shouldSwap(parentIndex, childIndex)) {
+      this._swap(parentIndex, childIndex);
       parentIndex = childIndex;
       childIndex = this._compareChildrenOf(parentIndex);
     }
   }
 
   /**
+   * Removes and returns the root node in the heap
+   * @public
+   * @returns {object}
+   */
+  extractRoot() {
+    if (this.isEmpty()) return null;
+
+    const root = this.root();
+    this._nodes[0] = this._nodes[this.size() - 1];
+    this._nodes.pop();
+    this._heapifyDown();
+
+    if (root === this._leaf) {
+      this._leaf = this.root();
+    }
+
+    return root;
+  }
+
+  /**
+   * Pushes a node with down in the heap before an index
    * @private
-   * pushes the swapped node with root down in its correct location
-   * @param {number} i -  swapped node's index
+   * @param {number} index
    */
   _heapifyDownUntil(index) {
     let parentIndex = 0;
@@ -136,8 +243,8 @@ class Heap {
         rightChildIndex
       );
 
-      if (this._shouldSwap(childIndex, parentIndex)) {
-        this._swap(childIndex, parentIndex);
+      if (this._shouldSwap(parentIndex, childIndex)) {
+        this._swap(parentIndex, childIndex);
       }
 
       parentIndex = childIndex;
@@ -147,22 +254,22 @@ class Heap {
   }
 
   /**
+   * Returns a shallow copy of the heap
    * @protected
-   * returns a shallow copy of a heap
    * @param {class} HeapType
    * @returns {Heap}
    */
   _clone(HeapType) {
-    return new HeapType(this._nodes.slice());
+    return new HeapType(this._nodes.slice(), this._leaf);
   }
 
   /**
+   * Sorts the heap by swapping root with all nodes and fixing positions
    * @public
-   * implements heap sort algorithm by swapping root with i nodes
    * @returns {array} the sorted nodes
    */
   sort() {
-    for (let i = this._getLastIndex(); i > 0; i -= 1) {
+    for (let i = this.size() - 1; i > 0; i -= 1) {
       this._swap(0, i);
       this._heapifyDownUntil(i);
     }
@@ -171,23 +278,62 @@ class Heap {
   }
 
   /**
+   * Inserts a node in the right position into the heap
    * @public
-   * inserts a node into the heap
    * @param {number|string} key
-   * @param {object} value
-   * @returns {HeapNode}
+   * @param {any} [value]
+   * @returns {Heap}
    */
   insert(key, value) {
-    const newNode = new HeapNode(key, value);
+    const newNode = value !== undefined ? { key, value } : key;
     this._nodes.push(newNode);
-    this._heapifyUp();
-    return newNode;
+    this.heapifyUp();
+    if (this._leaf === null || !this._compare(newNode, this._leaf)) {
+      this._leaf = newNode;
+    }
+    return this;
   }
 
   /**
+   * Fixes all positions of the nodes in the heap
    * @public
-   * returns the root node in the heap
-   * @returns {HeapNode}
+   * @returns {Heap}
+   */
+  fix() {
+    for (let i = 0; i < this.size(); i += 1) {
+      this.heapifyUp(i);
+    }
+    return this;
+  }
+
+  /**
+   * Verifies that the heap is valid
+   * @public
+   * @returns {boolean}
+   */
+  isValid(parentIndex = 0) {
+    let isValidLeft = true;
+    let isValidRight = true;
+
+    if (this._hasLeftChild(parentIndex)) {
+      const leftChildIndex = this._getLeftChildIndex(parentIndex);
+      if (!this._compareByIndex(parentIndex, leftChildIndex)) return false;
+      isValidLeft = this.isValid(leftChildIndex);
+    }
+
+    if (this._hasRightChild(parentIndex)) {
+      const rightChildIndex = this._getRightChildIndex(parentIndex);
+      if (!this._compareByIndex(parentIndex, rightChildIndex)) return false;
+      isValidRight = this.isValid(rightChildIndex);
+    }
+
+    return isValidLeft && isValidRight;
+  }
+
+  /**
+   * Returns the root node in the heap
+   * @public
+   * @returns {object|number|string|null}
    */
   root() {
     if (this.isEmpty()) return null;
@@ -195,41 +341,17 @@ class Heap {
   }
 
   /**
+   * Returns a leaf node in the heap
    * @public
-   * returns a leaf node in the heap
-   * @returns {HeapNode}
+   * @returns {object|number|string|null}
    */
   leaf() {
     return this._leaf;
   }
 
   /**
+   * Returns the number of nodes in the heap
    * @public
-   * removes and returns the root node in the heap
-   * @returns {HeapNode}
-   */
-  extractRoot() {
-    if (this.isEmpty()) return null;
-
-    const root = this.root();
-    this._nodes[0] = this._nodes[this._getLastIndex()];
-    this._nodes.pop();
-    this._heapifyDown();
-
-    if (root === this._leaf) {
-      if (this.isEmpty()) {
-        this._leaf = null;
-      } else {
-        this._leaf = this.root();
-      }
-    }
-
-    return root;
-  }
-
-  /**
-   * @public
-   * returns the number of nodes in the heap
    * @returns {number}
    */
   size() {
@@ -237,8 +359,8 @@ class Heap {
   }
 
   /**
+   * Checks if the heap is empty
    * @public
-   * checks if the heap is empty
    * @returns {boolean}
    */
   isEmpty() {
@@ -246,8 +368,8 @@ class Heap {
   }
 
   /**
+   * Clears the heap
    * @public
-   * clears the heap
    */
   clear() {
     this._nodes = [];
@@ -255,27 +377,31 @@ class Heap {
   }
 
   /**
+   * Convert a list of items into a heap
    * @protected
    * @static
-   * convert a list of items into a heap
-   * @param {array} items
+   * @param {array} array
    * @param {class} HeapType
    * @returns {Heap}
    */
-  static _heapify(items, HeapType) {
-    if (!isNoneEmptyArray(items)) return null;
+  static _heapify(list, HeapType) {
+    if (!Array.isArray(list)) {
+      throw new Error('.heapify expects an array');
+    }
 
-    const heap = new HeapType();
-    items.forEach((item) => {
-      if (isNumber(item) || isNoneEmptyString(item)) {
-        heap.insert(item);
-      } else if (isNoneNullObject(item)
-        && (isNumber(item.key) || isNoneEmptyString(item.key))) {
-        heap.insert(item.key, item.value);
-      }
-    });
+    return new HeapType(list).fix();
+  }
 
-    return heap;
+  /**
+   * Checks if a list of items is a valid heap
+   * @protected
+   * @static
+   * @param {array} array
+   * @param {class} HeapType
+   * @returns {boolean}
+   */
+  static _isHeapified(list, HeapType) {
+    return new HeapType(list).isValid();
   }
 }